Add files via upload
Foods are provided at the several co-ordinates of the screen for the snake to eat. Every time the snake eats the food, its length will by increased by one element along with the score.

The source code for Snake Game in C is complete and totally error-free. It is compiled in DEV C using the gcc compiler.